# [ITensors] Faster QN addition and filling