# Description

The main benefit of the `@inbounds` is that it lets Julia allocate `ma` on the stack instead of the heap.

A performance comparison:

```julia
using ITensors, BenchmarkTools
@benchmark qn + qn setup=(qn=QN("a" =>  1))
```

Performance before:
```
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  28.614 ns … 269.277 ns  ┊ GC (min … max): 0.00% … 72.18%
 Time  (median):     30.924 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   34.880 ns ±  15.130 ns  ┊ GC (mean ± σ):  2.09% ±  4.88%

 Memory estimate: 208 bytes, allocs estimate: 1.
```

Performance after:
```
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  23.370 ns … 80.040 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     23.571 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   24.034 ns ±  2.391 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

 Memory estimate: 0 bytes, allocs estimate: 0.
```

This along with the other change have real impact on the runtime of my MPO construction algorithm.


# Checklist:

- [x] My code follows the style guidelines of this project. Please run `using JuliaFormatter; format(".")` in the base directory of the repository (`~/.julia/dev/ITensors`) to format your code according to our style guidelines.
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand areas.
- [x] I have added tests that verify the behavior of the changes I made.
- [x] I have made corresponding changes to the documentation.
- [x] My changes generate no new warnings.
- [x] Any dependent changes have been merged and published in downstream modules.